+++ lib/ioc/component/decorators/index.ts
@@ -3,3 +3,5 @@ export * from './Implements';
 export * from '../componentUtils';
 export * from './Strategy';
 export * from './PostConstruct';
+export * from './OnStart';
+export * from './OnStop';