+++ lib/server/types/index.ts
@@ -1,2 +1,3 @@
 export * from './Class';
+export * from './ShutdownOptions';
 export * from '../web/middleware/AsenaStaticServeService';